# govctl_cli/api/openapi.yaml
openapi: "3.0.3"
info:
  title: "SoloCorp OS — Admin API"
  description: |
    REST API for SoloCorp OS governance, agent orchestration, pipeline
    monitoring, and system health. Used by the Dashboard UI and external
    integrations.

    **Base URL:** `http://localhost:8765`

    **Authentication:** API key via `X-API-Key` header (required for write ops).
  version: "1.0.0"
  contact:
    name: "Software Architect (พี่ทรงศักดิ์)"
    url: "https://github.com/DrSoloDev/Lab-solocorp-os2.4"
servers:
  - url: http://localhost:8765
    description: "Local development server"
  - url: https://api.solocorp.local
    description: "Production server (future)"
paths:
...
  /api/v1/gov/adrs:
    get:
      operationId: list_adrs
      summary: List Architecture Decision Records
      description: |
        Returns all ADRs from the governance directory. Supports optional
        filtering by status, domain, and search across title/context.
      tags: [Governance]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [proposed, accepted, deprecated, superseded]
          description: "Filter by ADR status"
          required: false
        - name: domain
          in: query
          schema:
            type: string
          description: "Filter by governance domain (e.g. architecture, security)"
          required: false
        - name: search
          in: query
          schema:
            type: string
          description: "Search in title and context text"
          required: false
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: "Maximum results to return"
          required: false
      responses:
        "200":
          description: "List of ADRs"
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ADRSummary"
                  total:
                    type: integer
                    description: "Total count matching filters (before limit)"
                  page:
                    type: integer
                    default: 1
        "400":
          $ref: "#/components/responses/BadRequest"

    post:
      operationId: create_adr
      summary: Create an Architecture Decision Record
      description: |
        Creates a new ADR artifact as a TOML file in the governance directory
        and publishes a governance event to Central Bus for audit trailing.
      tags: [Governance]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ADRCreateRequest"
      responses:
        "201":
          description: "ADR created successfully"
          headers:
            Location:
              schema:
                type: string
              description: "URL of the created ADR"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ADRDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/gov/adrs/{id}:
    get:
      operationId: get_adr
      summary: Get ADR detail
      description: "Returns the full ADR content including all metadata fields."
      tags: [Governance]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^ADR-\\d+$"
          description: "ADR identifier (e.g. ADR-001)"
      responses:
        "200":
          description: "ADR detail"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ADRDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===================================================================
  # Governance — RFC
  # ===================================================================
  /api/v1/gov/rfcs:
    get:
      operationId: list_rfcs
      summary: List Request for Comments
      description: |
        Returns all RFCs from the governance directory with optional filtering.
      tags: [Governance]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, review, approved, rejected, implemented]
          description: "Filter by RFC status"
          required: false
        - name: domain
          in: query
          schema:
            type: string
          description: "Filter by governance domain"
          required: false
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
      responses:
        "200":
          description: "List of RFCs"
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/RFCSummary"
                  total:
                    type: integer

    post:
      operationId: create_rfc
      summary: Create a Request for Comments
      description: |
        Creates a new RFC artifact as a TOML file and publishes a governance
        event to the Central Bus. The RFC triggers notification to relevant
        departments based on domain classification.
      tags: [Governance]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RFCCreateRequest"
      responses:
        "201":
          description: "RFC created successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RFCDetail"
  /api/v1/gov/rfcs/{id}:
    get:
      operationId: get_rfc
      summary: Get RFC detail
      tags: [Governance]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^RFC-\\d+$"
          description: "RFC identifier (e.g. RFC-001)"
      responses:
        "200":
          description: "RFC detail"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RFCDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===================================================================
  # Governance — Guards
  # ===================================================================
  /api/v1/gov/guards:
    get:
      operationId: list_guards
      summary: List verification guards
      description: |
        Returns all guard definitions and their current status across projects.
      tags: [Governance]
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
          description: "Filter by project"
          required: false
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, passed, failed]
          required: false
      responses:
        "200":
          description: "List of guards"
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/GuardSummary"
                  total:
                    type: integer

    post:
      operationId: create_guard
      summary: Create a guard check
      description: |
        Creates a new guard check for a project. This adds the guard to the
        project's active_guards list and sets guard_status to 'pending'.
        The guard must pass before the phase can transition to 'done'.
      tags: [Governance]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuardCreateRequest"
      responses:
        "201":
          description: "Guard created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuardDetail"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ===================================================================
  # Agent
  # ===================================================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key for write operations. Configurable in gov/config.toml"
  schemas:
...
    ADRSummary:
      type: object
      description: "Summary of an Architecture Decision Record"
      required: [id, title, status]
      properties:
        id:
          type: string
          example: "ADR-001"
        title:
          type: string
          example: "Use Postgres for event store"
        status:
          type: string
          enum: [proposed, accepted, deprecated, superseded]
        domain:
          type: string
          example: "architecture"
        impact:
          type: string
          example: "high"
        date:
          type: string
          format: date
        author:
          type: string
    ADRDetail:
      type: object
      description: "Full ADR content"
      allOf:
        - $ref: "#/components/schemas/ADRSummary"
        - type: object
          properties:
            context:
              type: string
              description: "The issue motivating this decision"
            decision:
              type: string
              description: "The change being proposed"
            consequences:
              type: string
              description: "What becomes easier or harder"
            options:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  description:
                    type: string
                  pros:
                    type: array
                    items:
                      type: string
                  cons:
                    type: array
                    items:
                      type: string
            path:
              type: string
              description: "Absolute path to the TOML file"
    ADRCreateRequest:
      type: object
      required: [title, status]
      properties:
        title:
          type: string
          description: "ADR title"
          example: "Use Postgres for event store"
        status:
          type: string
          enum: [proposed, accepted, deprecated, superseded]
          default: proposed
        context:
          type: string
          description: "The issue motivating this decision"
        decision:
          type: string
          description: "The change being proposed"
        consequences:
          type: string
          description: "What becomes easier or harder"
        domain:
          type: string
          example: "architecture"
        impact:
          type: string
          enum: [low, medium, high, critical]
          default: medium

    # ── RFC ──────────────────────────────────────────────────────────
    RFCSummary:
      type: object
      description: "Summary of a Request for Comments"
      required: [id, title, status]
      properties:
        id:
          type: string
          example: "RFC-001"
        title:
          type: string
          example: "Complexity Threshold Model"
        status:
          type: string
          enum: [draft, review, approved, rejected, implemented]
        domain:
          type: string
          example: "governance"
        author:
          type: string
        date:
          type: string
          format: date
        complexity_score:
          type: integer
          minimum: 0
          maximum: 3
    RFCDetail:
      type: object
      description: "Full RFC content"
      allOf:
        - $ref: "#/components/schemas/RFCSummary"
        - type: object
          properties:
            background:
              type: string
              description: "Background and motivation"
            proposal:
              type: string
              description: "The proposed change"
            alternatives:
              type: array
              items:
                type: string
            implementation:
              type: string
              description: "Implementation notes"
            path:
              type: string
    RFCCreateRequest:
      type: object
      required: [title, status]
      properties:
        title:
          type: string
          description: "RFC title"
        status:
          type: string
          enum: [draft, review, approved, rejected, implemented]
          default: draft
        background:
          type: string
        proposal:
          type: string
        domain:
          type: string
          example: "governance"
        complexity_score:
          type: integer
          minimum: 0
          maximum: 3
          default: 0

    # ── Guard ────────────────────────────────────────────────────────
    GuardSummary:
      type: object
      description: "Summary of a verification guard"
      required: [name, status]
      properties:
        name:
          type: string
          example: "adr_format"
        status:
          type: string
          enum: [pending, running, passed, failed]
        project_id:
          type: string
          example: "PRJ-1"
        added_at:
          type: string
          format: date-time
    GuardDetail:
      type: object
      allOf:
        - $ref: "#/components/schemas/GuardSummary"
        - type: object
          properties:
            phase:
              type: string
              description: "The phase this guard blocks"
            description:
              type: string
            resolved_at:
              type: string
              format: date-time
              nullable: true
            detail:
              type: object
              description: "Guard-specific result data"
    GuardCreateRequest:
      type: object
      required: [name, project_id]
      properties:
        name:
          type: string
          description: "Guard name (unique per project)"
        project_id:
          type: string
          description: "Project to attach the guard to"
        phase:
          type: string
          description: "Phase this guard applies to (optional)"
        description:
          type: string
          description: "Human-readable description"

    # ── Agent ────────────────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: "Invalid request parameters"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: object
    NotFound:
      description: "Resource not found"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: "RESOURCE_NOT_FOUND"
                  message:
                    type: string
    ValidationError:
      description: "Validation error"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: "VALIDATION_ERROR"
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                        message:
                          type: string
...